CSV file
I am using Steel Energy Dataset uploaded by Csafrit in Kaggle.
This company produces several types of coils, steel plates, and iron plates. The information on electricity consumption is held in a cloud-based system. The information on energy consumption of the industry is stored on the website of the Korea Electric Power Corporation (pccs.kepco.go.kr).